iFix issue #4636 F10 selects and copies
Improvement: Modify the function assigned to NVDA + F9 and F10 to
allow selecting the text as well as copying
https://github.com/nvaccess/nvda/issues/4636

Modified the existing NVDA+F9 then NVDA+F10 behaviour to
select text on the first press of F10, and copy to the
clipboard on the second press of F10.
If the user moves the cursor between the first and second
press of F10, the end marker is updated and the new text
range is selected.
The text will be copied if the end marker is the same
for both pushes of F10.